# backend/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Cue Queue Integration API
  version: 2.0.0
  description: |
    Cue is a hosted, multi-tenant queue integration service for banks.

    There are three authentication surfaces:

    * Public guest routes under `/api/v1` do not require authentication.
    * Dashboard routes under `/api` use a JWT Bearer token.
    * Authenticated v1 routes use an `X-API-Key` header, are rate-limited per key,
      and are restricted to the key's bank and scopes.

    The public v1 guest routes are mounted before the authenticated v1 pipeline.
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://api.example.com
    description: Production deployment
...
paths:
...
  /api/auth/login/admin:
    post:
      tags: [Auth]
      summary: Log in as bank admin
      security: []
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/auth/register/admin:
    post:
      tags: [Auth]
      summary: Register a bank admin
      description: Public self-registration. The bank name is required and the endpoint is rate-limited.
      security: []
      responses:
        '201': { $ref: '#/components/responses/Created' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
...
  /api/admin/api-key-requests:
    post:
      tags: [Admin]
      summary: Request a bank API key
      x-auth: jwt
      x-required-role: admin
      security:
        - bearerAuth: []
      responses:
        '201': { $ref: '#/components/responses/Created' }
        '409': { $ref: '#/components/responses/Conflict' }
    get:
      tags: [Admin]
      summary: List this bank's API key requests
      x-auth: jwt
      x-required-role: admin
      security:
        - bearerAuth: []
      responses:
        '200': { $ref: '#/components/responses/Success' }
  /api/admin/api-key-requests/{id}:
    get:
      tags: [Admin]
      summary: Get an API key request and reveal its raw key once
      x-auth: jwt
      x-required-role: admin
      security:
        - bearerAuth: []
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/platform/api-keys:
    post:
      tags: [Platform]
      summary: Create an API key
      x-auth: jwt
      x-required-role: superadmin
      security:
        - bearerAuth: []
      responses:
        '201': { $ref: '#/components/responses/Created' }
    get:
      tags: [Platform]
      summary: List platform API keys
      x-auth: jwt
      x-required-role: superadmin
      security:
        - bearerAuth: []
      responses:
        '200': { $ref: '#/components/responses/Success' }
  /api/platform/api-keys/{id}:
    delete:
      tags: [Platform]
      summary: Revoke an API key
      x-auth: jwt
      x-required-role: superadmin
      security:
        - bearerAuth: []
      responses:
        '200': { $ref: '#/components/responses/Success' }
...
  /api/platform/key-requests:
    get:
      tags: [Platform]
      summary: List bank API-key requests
      x-auth: jwt
      x-required-role: superadmin
      security:
        - bearerAuth: []
      responses:
        '200': { $ref: '#/components/responses/Success' }
  /api/platform/key-requests/{id}:
    patch:
      tags: [Platform]
      summary: Approve or reject an API-key request
      x-auth: jwt
      x-required-role: superadmin
      security:
        - bearerAuth: []
      responses:
        '200': { $ref: '#/components/responses/Success' }
...
  /api/v1/admin:
    get:
      tags: [V1]
      summary: Get the bank admin overview
      x-auth: apiKey + staff identity
      x-required-scope: analytics:read
      security:
        - apiKeyAuth: []
      responses:
        '200': { $ref: '#/components/responses/Success' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Dashboard JWT containing id, kind, and role claims.
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Scoped bank integration key. Authenticated v1 routes also resolve a Staff identity using X-Staff-Id or the key's defaultStaffId.
  responses:
    Success:
      description: Request succeeded
      content:
        application/json:
          schema: { $ref: '#/components/schemas/SuccessEnvelope' }
    Created:
      description: Resource created
      content:
        application/json:
          schema: { $ref: '#/components/schemas/SuccessEnvelope' }
    BadRequest:
      description: Validation failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Unauthorized:
      description: Missing, invalid, expired, or revoked credentials
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Conflict:
      description: Resource state conflict
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  schemas:
    SuccessEnvelope:
      type: object
      properties:
        status: { type: string, example: success }
        message: { type: string }
        data: { nullable: true }
        meta: { type: object, nullable: true }
    ErrorEnvelope:
      type: object
      properties:
        status: { type: string, example: error }
        message: { type: string }
        errors:
          type: array
          nullable: true
          items: { type: string }
...